reduce code to files
