+++ Picture_bot/confidence_info/your_dir.py
@@ -1,0 +1,1 @@
+main_img_dir = "C:/Users/tramp/source/repos/PythonApplication1/PythonApplication1/photos/"
